refactor: :truck: `example_package_properties.py` -> `example_properties.py`
