+++ src/seedcase_sprout/core/example_properties.py
